Renamed funtcion helper files
