+++ src/engine/source/builder/register.hpp
@@ -24,8 +24,8 @@
 #include "opBuilderConditionReference.hpp"
 #include "opBuilderConditionValue.hpp"
 #include "opBuilderFileOutput.hpp"
-#include "OpBuilderHelperFilter.hpp"
-#include "OpBuilderHelperMap.hpp"
+#include "opBuilderHelperFilter.hpp"
+#include "opBuilderHelperMap.hpp"
 #include "opBuilderMap.hpp"
 #include "opBuilderMapReference.hpp"
 #include "opBuilderMapValue.hpp"
