+++ src/engine/test/source/builder/builders/opBuilderHelperRegexMatch_test.cpp
@@ -11,7 +11,7 @@
 #include <vector>
 #include <re2/re2.h>
 
-#include "OpBuilderHelperFilter.hpp"
+#include "opBuilderHelperFilter.hpp"
 #include "testUtils.hpp"
 
 using namespace builder::internals::builders;
