+++ src/engine/test/source/builder/builders/stageBuilderCheckTest.cpp
@@ -14,7 +14,7 @@
 #include "opBuilderCondition.hpp"
 #include "opBuilderConditionReference.hpp"
 #include "opBuilderConditionValue.hpp"
-#include "OpBuilderHelperFilter.hpp"
+#include "opBuilderHelperFilter.hpp"
 #include "stageBuilderCheck.hpp"
 
 using namespace builder::internals::builders;
